Replace `dh-buildinfo` without the `dh_buildinfo` to ease transition
As requested in #1068809, `debhelper` will temporarily provide
`dh-buildinfo` with relevant `Conflicts` and `Replaces` to ensure
apt can find a solution. The provides is only on a metadata level,
since most packages no longer use `dh_buildinfo` directly or cope
gracefully with its removal. A few packages are expected to break,
which will be cleaned up by people who volunteered in #1068809.

If your package starts to FTBFS as a consequence of this change,
please remove the call to `dh_buildinfo`.

See: #1068809
Gbp-Dch: full